Big changes in this commit. Moved website dev to root, updating major portions of the informational site. Components page updated, along with more CSS rules and site design edits. Garbage clean up.
This brings my big changes front and center for the world to see. The informational site is more than halfway finished, although it still needs quite a bit of work. This is part of the "Chores" that need to be done before I can continue any significant work on Modevious.