Made the download page modular.
If JS is enabled the download section will show up instead of navigating to another page. If JS were to be disabled, for whatever reason, it will by default go to the original download page. No harm, graceful degradation. This is done mostly for fun, just for kicks. Also a decent example of the creative things one can do.