Created bash script for automatic compilation. Updated images on home page in dev.
The bash script will work on Linux and should work on Mac as well. Next up is the Windows version, which requires a few tweaks but the logic is pretty much all there. Git is already set up to ignore files in the build folder! Basically once you run the script the files will be created but will not be tracked by Git. This was made possible by one of my earlier commits.

Pivotal Tracker story:
http://www.pivotaltracker.com/story/show/5132480